Fix unit tests
This commit fixes up the unit test directory a bit so they can run
manually (assuming you have all the proper dependencies installed,
that'll be fixed in another patch). This just cleans up the tree so all
the tests are importable by unittest discovery, it removes the manual
test suite building which was out of date. It also fixes some pathing
errors which were assuming the tests were being run from the directory
they live in.