Enforce and fix lint on test code
We are currently only enforcing the lint and style checkers on the main
project code and ignoring the test code. We really should ensure a
consistent style throughout the entire repo if we're going to do it at
all. This commit does just that and adds lint enforcement for the tests.
In order to make this work it also fixes all the issues found in the
tests.