Add names to the travis jobs to better identify them
Right now when checking the status of a travis run the jobs aren't
identified by a name. You can figure out what each job is based on the
python version and the environment variables set. But, to make it a bit
easier this commit adds a name so each job is uniquely labeled and more
easily identifiable.