Add OSX jobs and stage travis
We support running ignis on windows, osx, and linux but we only
currently test running things on linux. This commit adds the travis
config for running tests under osx too so we can verify that it
works as expected. Since this basically doubles the number of travis
jobs we're running to conserve throughput this also adds job staging.
So when a PR is pushed first it will run the lint checks and python 3.5
tests on linux. If those pass then it will run tests unittests in the
remaining linux environments and on osx.