Disable python 3.7 and 3.5 jobs and remove conda usage
This commit removes the python 3.7 and 3.5 jobs because the aer wheels
published on pypi are broken and unusable. There is also no sdist
published to automate a build via pip. Cloning and building Aer from
source isn't worth it just for unit testing python 3.7 and 3.5 on
windows. So until there are working Aer builds published on pypi this
commit just comments out those jobs. We can add them back when/if there
are working windows wheels for python 3.7 and 3.5.

This commit also removes the conda usage since all that manipulation was
in an attempt to get the wheels to load against potentially missing
dynamic libs. But since that never worked there should be no need to
interact with conda (or use it) in the jobs. This comments them out just
in case we need to add it in the future.