Add appveyor config to test on windows (#151)
* Add appveyor config to test on windows

We support running ignis on windows, osx, and linux but we only
currently test running things on linux. This commit adds the initial
config for running tests under windows too so we can verify that it
works as expected.

* Switch to conda and install mkl

* Add -y to conda install

* Don't upgrade pip

* Tweak install steps slightly

* Tweak conda install command

* More conda tweaks

* Drop --user pip flag

* Re-add pip update pip

* Re-add pip update pip

* Switch to lists of file content instead of using filecmp in clifford tests

* Uncomment aer requirement

* Fix lint

* Install more via conda

* Fix typos and reorder

* No conda numpy

* Downgrade python 3.5

* Disable python 3.7 and 3.5 jobs and remove conda usage

This commit removes the python 3.7 and 3.5 jobs because the aer wheels
published on pypi are broken and unusable. There is also no sdist
published to automate a build via pip. Cloning and building Aer from
source isn't worth it just for unit testing python 3.7 and 3.5 on
windows. So until there are working Aer builds published on pypi this
commit just comments out those jobs. We can add them back when/if there
are working windows wheels for python 3.7 and 3.5.

This commit also removes the conda usage since all that manipulation was
in an attempt to get the wheels to load against potentially missing
dynamic libs. But since that never worked there should be no need to
interact with conda (or use it) in the jobs. This comments them out just
in case we need to add it in the future.

* Switch back to conda to install mkl for cvx