Attempt to unblock CI by running Aer from source
Things in CI are wedged right now because of a qiskit-terra master
incompatibility with the released version of aer, 0.11. There is a patch
up to qiskit-terra, Qiskit/qiskit-terra#2154, that should fix that but
it's blocked by another issue with aer. In an attempt to unblock CI
until that patch is merged this commit switches to run aer from master.
However, because of test environment mismatch this should be reverted as
soon as Qiskit/qiskit-terra#2154 merges.