Bump version number in setup.py
After we released 0.1.0 we should have bumped the version number in the
package to indicate that master is now ahead of the 0.1.0 version (since
that's in the past). However we never did that and when installing ignis
from master it gets evaluated as the same version as 0.1.0. This commit
corrects the oversight and bumps the version.