Stop building aer from source
Building aer from source is very slow, unreliable, and doesn't work on
windows at all. It also requires a lot of hacks to even work in the CI
system. This was only needed because terra master (soon to be 0.8)
didn't work with the latest release of aer (0.1.1). However, now that
Qiskit/qiskit-terra#2154 has merged we should be able to use master
terra with the released version of Aer. So we can stop building it from
source and speed up and improve our jobs.