Try to re-enable python35 and python37 jobs on appveyor
Now that we have a new aer release with new prebuilt binaries hosted on
pypi let see if they work in the windows CI environment on appveyor now.
Previously we disabled the windows testing because the binaries
published for aer 0.1.1 did not work, but if the new version works it'll
be good to verify that ignis runs on all 3 supported version of python
on windows.