Don't install terra from master in tests
Now that terra 0.8.x is out ignis doesn't have any requirements on terra
>0.8.x now. However it's not keeping up with the breaking changes being
made in terra's master branch. For the time being remove the step
install terra from master to avoid breakages by terra development.
When/if we need to ensure compatibility with a newer version of terra we
can bump the minimum version and add terra from master back to the
test environments.