Move CONTRIBUTING.md to top level (#216)
The CONTRIBUTING.md doc contains a lot of useful information about the
not only contributing to the project but also some useful information
about project policy and conventions. But it's currently in a hidden
directory so that github ui will show it on a help menu. So unless you
know to click on that help menu or that it's hiding in the .github
directory you'll never see it. This commit addresses this by moving the
file to the repo root so it's more obvious where it is.

Mirror of Qiskit/qiskit-terra#2359 and Qiskit/qiskit-aer#229