Move CODE_OF_CONDUCT to top level (#217)
The CODE_OF_CONDUCT.md doc contains information about the project and
community that we want to be clearly discoverable by people. But it's
currently in a hidden directory so that github ui will show it on a help
menu. So unless you know to click on that help menu or that it's hiding
in the .github directory you'll never see it. This commit addresses this
by moving the file to the repo root so it's more obvious where it is.
This is a continuation from #216 since it merged before this could be
added there.