Add version string to ignis
Unlike the other elements (and most python packages) ignis does not set
a `__version__` string for the root of it's package. This makes it
difficult to determing which version of ignis is being used at run time
from inside python. This commit fixes this by adding a version string to
ignis. It also will properly report if you're running a release or a dev
version (including a short sha1).

As part of this the setup.py is reworked to properly list the packages
in ignis using a feature in a newer setuptools (instead of a custom
function which didn't handle files in the ignis namespace root
correctly). Since this wasn't a requirement before the lack of the
newer setuptools is caught and reported to the user instead of failing
with an AttributeError during the install.