Update the mailmap (#226)
* Update the mailmap

This commit updates the mailmap to actually update names and email
address. Before it was just listed as a desired end state but without
actually containing the map (ie what we change from). This commit
updates the entries in the mailmap file to be a valid mapping with
the desired output. If the log entries are already in that format it
removes them from the mailmap because there is no benefit to mapping to
the same thing. It also updates the comment to be a bit more clear on
how to use the mailmap file.

* Mailmap consistency with terra