Add pypi automation for releases (#234)
As part of streamlining the release process, this commit adds automation
for publishing to pypi. For the metapackage this is straightforward
since we only publish an sdist right now and there is no code we just
need to build an sdist and a python3 wheel and then upload both to pypi.
This takes another manual step out of the release process and all that
is necessary is to push a tag (or use github's ui for creating a tag
and/or release).