Make azure pipeline steps fail on intermediate commands (#248)
* Make azure pipeline steps fail on intermediate commands

The azure pipelines commands were previously using the script directive
which on windows uses cmd.exe to run commands. The return code for the
step process only returns non-zero and exits if the last command in the
script returns non-zero, it will continue running if an intermediate
command fails. This can lead to hard to parse errors because something
failed earlier in the pipeline and we don't catch it until later. For
example pip install tox failing because of a network issue and then we
get a command not found error for tox later when it goes to run tests.
This commit fixes this potential issue by switching all the script to
use base and then using 'set -e' to set errexit to make sure any failing
command causes the job to fail there.

* Use bash style variables instead of cmd

* Add trace mode for debugging