Revert "Make azure pipeline steps fail on intermediate commands (#248)" (#249)
This PR was merged by mistake before it was ready. The switch to bash
scripts breaks the anaconda commands. It still needs to be debugged.
Eventually we'll want this change because it'll make our jobs more
reliable in the long run. But the last details still need to be worked
out. This commit reverts the change to unblock the windows CI until it
is ready.

This reverts commit e3420a6f88bc1e345cf4ba349545d290785d549e.