Execute tests in parallel for CI
This commit switches the test runner we use to use stestr [1] for
running tests in parallel instead of the built in unittest test runner
which just runs things serially. Right now a test job takes ~3 minutes
to run all the tests serially. This can be sped up by running tests in
parallel. In testing locally it's consistently taking between 70-80
secs with parallel execution.

[1] https://stestr.readthedocs.io/en/latest/MANUAL.html