Add caching of stestr timing data
This commit adds caching of the stestr timing data so that the scheduler
can use that data to improve the test run times. The only catch is that
we need to ensure that travis doesn't create a blank directory when
loading from the cache because that will break the test execution.