Pin osqp version for tests (#264)
osqp gets pulled in as a dependency by cvxpy but the recent 0.6.0 (and
0.6.1) releases on pypi did not include a prebuilt binary wheel for
python 3.5 on Windows and macOS, see: oxfordcontrol/osqp#187. This
breaks our windows CI on python 3.5 because we do not have the
environment configured to install that package from source. To work
around this in the meantime this commit pins the version of osqp we
install to be the previous release, 0.5.0, which had prebuilt binaries
for all our testing environments.