Update the CONTRIBUTING.md file (#266)
The contributing doc was quite out of date. It looks like it was
originally copy and pasted from an old version of the terra CONTRIBUTING
file and contained information that was never accurate for ignis. It
also had a number of markdown style issue and a lot of language that was
worded in a confusing manner. This commit updates all of this and tries
to make the document up to date and easier to read.

Fixes #230