Execute tests in parallel for CI (#265)
* Execute tests in parallel for CI

This commit switches the test runner we use to use stestr [1] for
running tests in parallel instead of the built in unittest test runner
which just runs things serially. Right now a test job takes ~3 minutes
to run all the tests serially. This can be sped up by running tests in
parallel. In testing locally it's consistently taking between 70-80
secs with parallel execution.

[1] https://stestr.readthedocs.io/en/latest/MANUAL.html

* Group tests by class

* Revert "Group tests by class"

This did not provide any speed up so there's no need to do any grouping.
Instead let the stestr scheduler individually split methods between
workers since that gives it more flexability in the local execution case
where there is historical timing data to inform scheduling decisions.

This reverts commit d3e68345e1c487c6bac3d2f8e38501efd5a2a0cb.

* Add caching of stestr timing data

This commit adds caching of the stestr timing data so that the scheduler
can use that data to improve the test run times. The only catch is that
we need to ensure that travis doesn't create a blank directory when
loading from the cache because that will break the test execution.