Split up test_rb using ddt
The test_rb test case is the slowest test and is built by having
nested loops to go over all the combinations of the number of qubits,
rb pattern, and multiplier type. This results in 2 issues, first the
test case is quite slow, but more importantly it's very difficult to
debug. By running the same test code over and over again with different
parameters it makes detecting a problem harder (especially when tests
run in parallel) but also it makes reproducing a failure much more
difficult because there is no way to run the test with just one set of
input parameters.

This commit addresses this by switching the test to use ddt instead. ddt
will let us reuse the same test code from inside the loops but instead
of doing it one giant test case it will construct separate smaller test
cases for each input parameter combination. This leads to faster
execution when running in parallel because the scheduler can split the
test up over multiple workers. But also gives us a separate test case
that can be rerun by itself when an error is encountered.