Fix lint failures caused by new pylint and pin
This commit fixes all the issues reported via the newest pylint release.
After fixing all the reported errors this also adds a version pin using
a constraints file to ensure that we have a stable base for lint
checking. We can investigate bumping that pin at the future on our
terms instead of new pylint (or pylint's dependencies) releases breaking
CI periodically.

At the same time the python version used for the lint check is bumped to
3.7. The python version used for pylint can effect results returned,
pylint is very sensitive to the environment it's run in. Running on
python 3.5 which is the oldest supported python version masks issues on
newer python versions. For example, if you have python 3.7 installed
locally lint will always fail becaues of changes to the stdlib. This
starts to become more of an issue as a particular python version ages,
since the user and contributor base moves to newer versions of python.
This commit changes the version of pylint we run in CI to 3.7 to tests
the leading edge instead of the trailing so we keep our pylint testing
up to date with the python versions we support.