[WIP] Discriminator plotting #273 (#278)
* * Added abstract plotting methods to BaseDiscrimination fitter.
* Added implementation of plotting methods in IQDiscriminationFitter
  that can
  - Plot decision boundaries
  - Plot training data
  - Flag misclassified training points
  - Plot data to be classified

* * Added matplotlib to the dependencies.
* Added a unit test for the plotting.

* * Changed an assertEquals to assertEqual.

* * Small change to test docstring.

* * Added qubits_to_plot in discriminator's plot method. Allows the
  user to plot a subset of qubits in qubit_mask.

* Added grouping of scatter plots by expect states.

* Added .idea to .gitignore.

* * Added functionality to add a title to each plot.
* Amended the example notebook.

* * Pickled a qiskit Result containing level 1 data that can be used
  for examples and as a test data set.

* Modified the example notebook. It can now obtain level 1 data
  either by running on the IBM Q hardware or by loading the test data.

* Fix lint failures caused by new pylint and pin (#279)

This commit fixes all the issues reported via the newest pylint release.
After fixing all the reported errors this also adds a version pin using
a constraints file to ensure that we have a stable base for lint
checking. We can investigate bumping that pin at the future on our
terms instead of new pylint (or pylint's dependencies) releases breaking
CI periodically.

At the same time the python version used for the lint check is bumped to
3.7. The python version used for pylint can effect results returned,
pylint is very sensitive to the environment it's run in. Running on
python 3.5 which is the oldest supported python version masks issues on
newer python versions. For example, if you have python 3.7 installed
locally lint will always fail becaues of changes to the stdlib. This
starts to become more of an issue as a particular python version ages,
since the user and contributor base moves to newer versions of python.
This commit changes the version of pylint we run in CI to 3.7 to tests
the leading edge instead of the trailing so we keep our pylint testing
up to date with the python versions we support.

* * Removed plotting test.

* Changed the way matplotlib is imported to conform with the rest of
  qiskit.

* * Added usage of HAS_MATPLOTLIB.

* Removed duplicate matplotlib import, also replace with try/except, remove from requirements-dev

* Updated the example with a flag to use pickled data, removed decorators from the functions that needed matplotlib

* fix lint