Add docs job to run docs build
Right now there are no jobs that check that the sphinx build using the
ignis docstrings actually succeed. This commit adds an azure pipelines
job to run the docs build. Azure is used instead of travis because it
includes an artifacts store for output from the job which we can dump
the built output in after a successful run. This will enable people to
download it and view the output locally. Additionally, a tox job,
tox -edocs, is added so reproducing the docs builds locally is
straightforward.