Fix sphinx warnings and enforce no warnings (#290)
PR #288 got a good start to cleaning up the docstrings in all the
modules in ignis. From that base the goal of fixing all the formatting
warnings from sphinx was much closer. This commit goes through all the
warnings reported by sphinx during a build and fixes the formatting in
the docstrings to avoid them. Then with a warning free build the -W flag
for sphinx-build is added by default to the build jobs to make all
sphinx warnings fatal in the future. While there still alot of things
that can be cleaned up in the docs having a warning free baseline will
make that all easier. The sphinx warnings while not fatal by default
almost always highlight real formatting issues that should be fixed.