Pin pywin32 version (#299)
Pywin32 was release on Nov. 10 and introduced an issue when running
virtualenv on windows, mhammond/pywin32#1439. This causes our ci system
to fail when trying to run the windows jobs because tox is uses
virtualenv underneath. This commit adds a pin via the constraints file
to ensure that we are installing the older version which does not have
this problem.