Set a consistent line length to 100 characters
This commit updates the pycodestyle rules used to match what is set in
pylint. Previously pycodesytle was using the default settings which has
an 80 character line length, but we set pylint to have a limit of 100.
This updates the pycodestyle configuration (via the tox.ini) to also
use 100 so they're consistent.