Bump macos version in travis
Currently we run macOS CI on macOS 10.12 which was released ~4 yrs ago
and end of life and no longer supported by Apple. Terra added a new
binary requirement retworkx which builds it's wheels against macOS
10.13. This means that pip can't find the wheels for macOS in CI
because 10.13 is too new. This commit bumps the macOS/xcode version
we use in the travis jobs to one slightly newer using macOS 10.13
(which is currently the default travis env). This should unblock CI
because we don't need to install retworkx from source.