Install git from master on pylint and docs jobs
This commit updates the tox job definitions for the lint checks and docs
builds to include terra from master. Right now these jobs were depending
on the requirements list for ignis to install terra as a dependency.
Which meant that we only ever terra installed from pypi. This was
blocking changes that depend on new features or other api changes from
terra. Ensuring that we install terra from master for all jobs ensures
that the test envs are all consistent.