Copy pylint config from terra (#370)
* Copy pylint config from terra

To make contribution to ignis more consistent with the other elements
this commit copies the pylint config used on terra over to ignis and
fixes the lint issues it identifies. The only difference between the 2
configs is the fileheader check is excluded in the ignis copy. This is
because the check isn't super reliable and also precludes running in
parallel.