Update CONTRIBUTING guide add reno for release notes (#396)
Moving forward ignis will be using reno for release documentation. This
commit adds a pro-forma release note and the documentation configuration
to verify release notes moving forward. It also updates the CONTRIBUTING
guide to reflect this new change. At the same time the majority of
contributing guide is rewritten since a large portion of the information
was duplicated with content in the combined qiskit documentation or out
of date.

Co-authored-by: Christopher J. Wood <cjwood@us.ibm.com>