Avoid implict marshmallow usage
The Results class in terra is no longer constructed via marshmallow
since #4030 merged. When updating the counts in a result object the
dictionary can now be set directly instead of needlessly wrapping it in
a marshmallow container for a dictionary. Besides fixing compatibility
with terra 0.15.0/master it also will improve performance because there
is no extra overhead from marshmallow constructing and running
validation prior to updating the attribute.