Unblock CI by updating jupyter_sphinx and pinning ddt
A recent release of jupyter-sphinx and ddt have caused CI failures. For
jupyter-sphinx the plugin path we were using has been deprecated which
causes a warning to be emitted, which is fatal in CI. Additionally a ddt
release added a requirement by not to the requirements list (see
datadriventests/ddt#83 for more details). This pins blacklists that ddt
version so it's not used. With these 2 fixes CI should be unblocked.