Skip tests that use broken pickle files
Several ignis unit tests have pickle files that are no longer valid
because they are using a now removed terra module. Until these can be
removed and the tests and ignis refactored to not rely on pickle files
(see #366 for more details) this commit just skips the problematic tests
because there is no way to regenerate these pickles or fix the
underlying tests currently.