Unblock CI by updating jupyter_sphinx and pinning ddt (#401)
* Unblock CI by updating jupyter_sphinx and pinning ddt

A recent release of jupyter-sphinx and ddt have caused CI failures. For
jupyter-sphinx the plugin path we were using has been deprecated which
causes a warning to be emitted, which is fatal in CI. Additionally a ddt
release added a requirement by not to the requirements list (see
datadriventests/ddt#83 for more details). This pins blacklists that ddt
version so it's not used. With these 2 fixes CI should be unblocked.

* Avoid implict marshmallow usage

The Results class in terra is no longer constructed via marshmallow
since #4030 merged. When updating the counts in a result object the
dictionary can now be set directly instead of needlessly wrapping it in
a marshmallow container for a dictionary. Besides fixing compatibility
with terra 0.15.0/master it also will improve performance because there
is no extra overhead from marshmallow constructing and running
validation prior to updating the attribute.

* Skip tests that use broken pickle files

Several ignis unit tests have pickle files that are no longer valid
because they are using a now removed terra module. Until these can be
removed and the tests and ignis refactored to not rely on pickle files
(see #366 for more details) this commit just skips the problematic tests
because there is no way to regenerate these pickles or fix the
underlying tests currently.