Add README as pypi long description
The long description field was missing from the pypi package this leads
to the package page on pypi showing "The author of this package has not
provided a project description" instead of any details about the
package. This commit updates the package metadata to use the README as
the long description so that a rendered version will be shown on the
pypi page.