Add extra requirements and document in readme
Ignis has 2 optional dependencies which are not required to use the
core functionality but are required to use certain features. These
weren't always clearly documented or exposed so it could potentially
come as a surprise to end users. This commit makes 3 changes around the
use of optional dependencies. First it updates the cvx error message to
more clearly explain the options. Then it adds an extras section to the
setup.py so that users can install the optional dependencies when they
install ignis. Finally a section is added to the readme to document the
optional dependencies.

Fixes #312