Make logging tests deterministic
The current iteration of the tests are constantly racing against each
other because they touch the same file paths. Tests are run in parallel
for speed reasons and there is no guarantee that 2 test methods aren't
running at the same time. This commit fixes these issues by using
tempfiles for the places the logger writes to disk. This way there is
never going to be 2 files being touched by different tests at the same
time.

The next step here would to replace all the file usage with mocks so we
don't actually write to disk. There is not really a reason to actually
write out to disk and it makes the tests more complicated because
we have to synchronizing the state of the FS or create a blank workspace
for each test. But this isn't done here because it would require a
slightly larger refactor of the tests.

Fixes #390