Reset in setUp too
The singleton nature of IgnisLogger class means that any logger state
initialized in a prior test will leak through to any follow on tests.
This breaks any tests of the logger because it assumes a clean state.
This commit adds a protective reset call to the the logger class setUp
to avoid this problem.