Revert #400/#401 from stable branch
The change made in #400 (which was missed because it got integrated into
the #401 PR which was needed to unblock CI on master) should not have
been backported. It was only necessary for compat with terra master and
terra 0.15.0 because the validation module is removed in 0.15.0. This is
breaking measurement mitigation on the stable branch and 0.3.1 release
which needs to work with terra 0.14.x. This commit reverts that change
so the we're wrapping the counts in qiskit.validation.Obj and the
Results object is useable in marshmallow with released terra.