Fix azure pipelines (#430)
In #422 the azure pipelines jobs were updated to simplify the dependency
installation to no longer require build scs against openblas. However
there was a bug in that which caused the jobs to not really run, but
still return success. This commit corrects the oversight so we're still
running CI for windows and actually testing it.