Fix breaking attribute name change
In #370 the name of the N_acc attribute was renamed to n_acc to adapt to
the pylint naming rules. However, this wwas neglecting that this
attribute was part of the API for the class, and people were using it as
N_acc. This commit reverts the breaking change and changes the name back
to N_acc.