Remove storaed data and run Aer during fitter test
Previously the results from executionn of qv circuits on Aer were run
once and stored in the repo. Then the fitters tests would just load the
stored results and fit that data. This was presumably done for run time,
however in practice it only takes a few seconds to simulate the circuits
and is not a significant overhead. This commit removes the generated
data and moves the circuit execution pieces from the generation script
into the test file.