Replace shallow copy with deepcopy in QV circuit generation
In #460 we migrated the QV circuit generation to use the QV circuit from
terra's circuit library instead of constructing it from scratch. As part
of that the circuit generation was updated to construct a single virtual
circuit via the library and just copy it for the no measure output
circuits. However, this was done using a shallow copy (in the interest
of performance) which means that any references between the circuits
will not be duplicated. For cases where the qubit_lists input is not a
contiguous list this would result in the measurements showing up in the
no measurement circuits because of shared references. This commit fixes
this by using deepcopy instead of a shallow copy so that there are no
shared references between the measurement circuit and the no measurement
circuit.