Use conda for windows CI
We need cvxopt in the windows jobs to have a functional solver for the
cvx fitter test (or manually build scs from source against blas).
However there are no wheels published for cvxopt on pypi for windows
python3.8. To workaround this we were using conda in azure pipelines,
this commit mirrors this by switching to conda for the windows tests.