Pin setuptools version (#485)
Setuptools 50.0.0 was released recently and since then this has been
blocking windows CI because it is causing scs dependency of cvxpy to
fail to install. This commit should fix this for the time being by
capping the setuptools version we use in CI until setuptools or scs
fixes the issues.