Migrate all CI to Github Actions (#476)
* Migrate all CI to Github Actions

Since the period with an extra 10 concurrent job on travis is ending and
there is already extensive resource contention in the shared azure
pipelines instance for the qiskit organization we need to find an
alternative for CI to maintain job throughput. This commit migrates all
the existing CI jobs to github actions which has a very generous quota
for open source projects.

* Fix release workflow

* Tweak trigger

* Get all history for docs job

* Install cvxopt in windows jobs

* Use bash on windows

* Use conda for windows CI

We need cvxopt in the windows jobs to have a functional solver for the
cvx fitter test (or manually build scs from source against blas).
However there are no wheels published for cvxopt on pypi for windows
python3.8. To workaround this we were using conda in azure pipelines,
this commit mirrors this by switching to conda for the windows tests.

* Upgrade setuptools and friend too

* Fix duplication

* Use conda forge

* Try conda installing scs too

* Only conda on windows 3.8

* Fix typo

* Actually install tox

* Upgrade setuptools on windows non-conda job

* Install cvxopt in tox venv on windows

* Fix more typos

* Fix / for bash

* Just use conda for all windows jobs

* Drop 3.5 jobs

* Auto update conda

* Test release job too

* Try adding 3.5 jobs back now that there is a working pattern

* Remove windows 3.5 job

* Fix syntax error in release job

* Tweak release artifact build step

* Ensure wheel is up to date for release job

* Upload release artifacts before twine push

* Only run wheel jobs on tags

* Disable parallelism for macOS 3.8 job

* Disable parallelism in macOS CI