Stop using conda in CI
The recent update of conda in the base images for windows CI has broken
the windows CI jobs because virtualenv is unable to discover the path to
the python binary. This commit attempts to fix this by abandoning the
conda usage and relying solely on system installed python and pip to
install the dependencies. The complexity here is that the SDP solver
installation has been problematic in windows outside of conda as cvxgrp
packages only seem to reliable ship binaries via conda for all
platforms.