Make scikit-learn an optional dependency
We previously had scikit-learn as a requirement that always needed to be
installed. However it was only ever used by the measurement
discriminators and not needed for ignis in general. Since this is a
fairly large package it is a bit heavy to require for every install
especially if the discriminators aren't being used. This commit removes
scikit-learn from the requirements list and adds it as an optional
dependency.