Fix ignis without optional requirements installed (#533)
* Fix ignis without optional requirements installed

In #500 type hints were added for the plotting functions using the
optional matplotlib dependencies. This breaks ignis installations
without matplotlib present since it's an optional dependency. Since type
hints are really just for documentation and not necessary this commit
just removes those problematic hints and uses the docstring to document
the types.

At the same time the testing job verifying ignis works without any
optional dependencies that was lost in the migration from travis to
github actions is restored to verify that everything works moving
forward.