Update aer check in topological codes (#546)
In Qiskit/qiskit-terra#5619 how the qiskit.Aer provider gets loaded will
be changing and it will always be present, not just when Aer is
installed. In the topological codes fitters it was checking for the
presence of Aer using qiskit.Aer which won't continue to work moving
forward. To avoid this potential issue (and potentially unblock
Qiskit/qiskit-terra#5619) this commit switches to using
'from qiskit.providers.aer import Aer' instead which lives in the
qiskit-aer project itself and will always import error if aer isn't
installed.