Migrate topological codes fitters to use retworkx
This commit migrates the networkx usage in the topological codes fitters
module away from networkx to use retworkx instead. This should provide
a speedup as retworkx is typically much faster than networkx. But, more
importantly this module is the last usage of networkx in qiksit and that
causes noticeable import overhead as networkx is slow to import. Moving
to retworkx like the rest of Qiskit will remove this overhead.