Prepare 0.6.0 release (#568)
* Prepare 0.6.0 release

To prepare for the upcoming 0.6.0 release this commit moves all the
release notes into a 0.6 subdirectory to keep them separate from future
development notes post-release. It also adds an incomplete release note
to add the prelude section providing the high level overview of the
release. When this merges it should be the commit used to tag the 0.6.0
release.

* Cleanup and fix release notes

* Fix expval bug note

* Remove operators tutorials from tutorial job too