Bump pylint version to the latest release
This commit bumps the pylint version we use for CI and local style
enforcement to be the latest release 2.8.2 and updates the code so that
the new pylint version passes. One place to note is a new rule looking
for potentially unbalanced tuple rule is disabled because the number of
values in the return tuple from the rb circuit generator varies
depending on the input (so we can't fix the code to pass the rule.
The version we were using was missing python 3.9 support and causes issues
for anyone running locally with python 3.9. This fixes this so we can
run pylint with all supported versions of python.