Update mailmap (#583)
This commit makes updates to the mailmap file to map emails and real names
from the git log to a canonical form. Most of the additions made here
are for regular qiskit contributors who were using the used multiple names
or email addresses and standardizes on the form already in a mailmap file
(either in terra or other qiskit repositories).