improve tests
The first version of these tests used `test_error_probs` as an end-to-end test, which should run for every type of code. This has now been changed to ensure better quality error messages, and shorter run times for the test. Though `test_error_probs` remains as a final confirmation, it only runs for the default version of the codes.

The bulk of the testing is now done by `test_graph_construction`, which reproduces the process used in the construction of the `GraphDecoder` object: every possible single qubit error is inserted and then the test ensures that this is detected in the required way. This is run on all types of code, for even and odd `d` and for one or two measurement rounds.