Add possibility of x base encoding (#581)
A keyword argument xbasis is added to RepetitionCodes. This allows an alternate encoding to be used, based on the repetitions of the X basis states rather than the Z basis.

This does not impact any existing features and adds no new methods. It basically just injects a bunch of hadamards and reverses cx gates if xbasis=True.

* add x basis encoding

* allow logical 1 analysis of errors

* change log to logical

* change uses of log to logical

* add documentation

* add documentation

* change xbasis encoding to false by default

* sped up prob calculation

* Revert "sped up prob calculation"

* make error prob calculation more efficient

* improve tests

The first version of these tests used `test_error_probs` as an end-to-end test, which should run for every type of code. This has now been changed to ensure better quality error messages, and shorter run times for the test. Though `test_error_probs` remains as a final confirmation, it only runs for the default version of the codes.

The bulk of the testing is now done by `test_graph_construction`, which reproduces the process used in the construction of the `GraphDecoder` object: every possible single qubit error is inserted and then the test ensures that this is detected in the required way. This is run on all types of code, for even and odd `d` and for one or two measurement rounds.

* Update test/topological_codes/test_codes.py

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* Update test/topological_codes/test_codes.py

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* add )

* linting

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>