Deprecate Qiskit Ignis
With the introduction of Qiskit Experiments development on Qiskit Ignis
has pretty much stopped. It's been in maintenance only mode for the past
6-8 months. Now that the only other use of Ignis inside Qiskit has been
worked around with Qiskit/qiskit-terra#6867 nothin depends on Qiskit Ignis
moving forward. This commit officially marks the deprecation of the project
and it will be retired and archived after the deprecation period
finishes.