Allow warnings in sphinx-build
Since importing ignis will raise a deprecation warning now this is
causing the sphinx builds to fail. Since we're not going to be
significantly changing the documentation prior to the retirement of
project it's easier to just allow warnings instead of filtering the
ignis deprecation warning only. This commit makes the change and removes
the -W flag from sphinx-build to unblock CI.