Unconditionally depend on jupyter_sphinx for dev
The original constraint was added in gh-505, at a time when Aer was
specifically requiring jupyter_sphinx in its own `requirements-dev.txt`
under certain constraints.  The constraint no longer appears to be
needed, so all it serves to do now is to cause the docs build to fail on
Python 3.9.