Drop 3.6 testing
Upstream Qiskit has dropped support for python 3.6. Since ignis is in
maintenance mode pending archive we can just drop testing for this. This
commit does this and removes the python 3.6 job. Since we're not doing
active feature development anymore we don't have to worry too much about
Python 3.6 compat. If need to release again before the archive we can
revisit the minimum supported python version in that release. As ignis
is pure Python there are less compatibility issues to worry about.