Drop 3.6 testing (#610)
* Drop 3.6 testing

Upstream Qiskit has dropped support for python 3.6. Since ignis is in
maintenance mode pending archive we can just drop testing for this. This
commit does this and removes the python 3.6 job. Since we're not doing
active feature development anymore we don't have to worry too much about
Python 3.6 compat. If need to release again before the archive we can
revisit the minimum supported python version in that release. As ignis
is pure Python there are less compatibility issues to worry about.

* Pin scs version

The latest scs releases are causing a failure. This commit pins it to
the last known working version in CI.

* Use windows server 2019 for ci jobs

* Pin cvxopt too

* Remove terra source install

* Drop broken windows CI env